# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QuizApp API Documentation",
        "contact": {
            "email": "haykalcs@gmail.com"
        },
        "version": "1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Register",
                "description": "Register User Here",
                "operationId": "authRegister",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "johndoe@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "Password`"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "example": "Password`"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "siswa",
                                            "guru"
                                        ],
                                        "example": "siswa"
                                    },
                                    "avatar": {
                                        "type": "file"
                                    },
                                    "number": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "nullable": [
                                    "avatar",
                                    "number"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registrasi berhasil"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "user": {
                                                    "name": "John Doe",
                                                    "email": "johndoe@gmail.com",
                                                    "email_verified_at": "2022-08-29T06:19:06.000000Z",
                                                    "role": "siswa",
                                                    "avatar": "1255762734.jpg",
                                                    "number": 2,
                                                    "created_at": "2022-08-29T06:19:06.000000Z",
                                                    "updated_at": "2022-08-29T08:34:52.000000Z",
                                                    "id": 8
                                                }
                                            }
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|pqLrmURrF0SmemD1TROWlWB1VIaUg1PE9r0uSf69"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "description": "Login User Here",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "siswa@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "Password`"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "siswa@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "Password`"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login berhasil"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "user": {
                                                    "id": 1,
                                                    "name": "siswa",
                                                    "email": "siswa@gmail.com",
                                                    "email_verified_at": "2022-08-29T06:19:06.000000Z",
                                                    "role": "siswa",
                                                    "avatar": null,
                                                    "number": null,
                                                    "last_seen": "2022-08-29T08:34:52.370167Z",
                                                    "created_at": "2022-08-29T06:19:06.000000Z",
                                                    "updated_at": "2022-08-29T08:34:52.000000Z"
                                                }
                                            }
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "2|pqLrmURrF0SmemD1TROWlWB1VIaUg1PE9r0uSf69"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Incorrect email or password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email atau password anda salah"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Logout",
                "description": "Logout User Here",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Logout success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout berhasil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}